requirements.txt
pandas
tqdm
numpy
pillow
imagehash
datasets
opencv-python
absl-py
ml-collections
huggingface-hub
gradio==4.44.1
pyperclip
# onnxruntime-gpu # required by Waifu Tagger
# onnxruntime # required by Waifu Tagger (cpu model only)
# torch # required by Waifu Scorer and Waifu Tagger
# torchvision
# git+https://github.com/openai/CLIP.git # required by Waifu Scorer